Major refactor to support legacy
