+++ jsdoc.conf.json
@@ -25,7 +25,7 @@
     },
     "markdown"  : {
         "parser"        : "gfm",
-        "hardwrap"      : true
+        "hardwrap"      : false
     },
     "opts": {
         "encoding"      : "utf8",
