+++ src/base/SoundSprite.ts
@@ -1,6 +1,6 @@
-import Sound from "./Sound";
-import {CompleteCallback} from "./Sound";
-import SoundInstance from "./SoundInstance";
+import BaseSound from "./BaseSound";
+import {CompleteCallback} from "./BaseSound";
+import {ISoundInstance} from "./ISoundInstance";
 
 export interface SoundSpriteData {
     start: number;
@@ -25,15 +25,15 @@ export default class SoundSprite
      * The reference sound
      * @name PIXI.sound.SoundSprite#parent
      * @type {PIXI.sound.Sound}
-     * @readOnly
+     * @readonly
      */
-    public parent: Sound;
+    public parent: BaseSound;
 
     /**
      * The starting location in seconds.
      * @name PIXI.sound.SoundSprite#start
      * @type {Number}
-     * @readOnly
+     * @readonly
      */
     public start: number;
 
@@ -41,7 +41,7 @@ export default class SoundSprite
      * The ending location in seconds
      * @name PIXI.sound.SoundSprite#end
      * @type {Number}
-     * @readOnly
+     * @readonly
      */
     public end: number;
 
@@ -49,7 +49,7 @@ export default class SoundSprite
      * The speed override where 1 is 100% speed playback.
      * @name PIXI.sound.SoundSprite#speed
      * @type {Number}
-     * @readOnly
+     * @readonly
      */
     public speed: number;
 
@@ -57,14 +57,14 @@ export default class SoundSprite
      * The duration of the sound in seconds.
      * @name PIXI.sound.SoundSprite#duration
      * @type {Number}
-     * @readOnly
+     * @readonly
      */
     public duration: number;
 
     /**
      * Constructor
      */
-    constructor(parent: Sound, options: SoundSpriteData)
+    constructor(parent: BaseSound, options: SoundSpriteData)
     {
         this.parent = parent;
         Object.assign(this, options);
@@ -81,7 +81,7 @@ export default class SoundSprite
      * @param {PIXI.sound.Sound~completeCallback} [complete] Function call when complete
      * @return {PIXI.sound.SoundInstance|Promise<PIXI.sound.SoundInstance>} Sound instance being played.
      */
-    public play(complete?: CompleteCallback): SoundInstance|Promise<SoundInstance>
+    public play(complete?: CompleteCallback): ISoundInstance|Promise<ISoundInstance>
     {
         return this.parent.play(Object.assign({
             complete,
