+++ src/filters/EqualizerFilter.ts
@@ -29,7 +29,7 @@ export default class EqualizerFilter extends Filter
      * Band at 32 Hz
      * @name PIXI.sound.filters.EqualizerFilter.F32
      * @type {Number}
-     * @readOnly
+     * @readonly
      */
     public static F32:number = 32;
 
@@ -37,7 +37,7 @@ export default class EqualizerFilter extends Filter
      * Band at 64 Hz
      * @name PIXI.sound.filters.EqualizerFilter.F64
      * @type {Number}
-     * @readOnly
+     * @readonly
      */
     public static F64:number = 64;
     
@@ -45,7 +45,7 @@ export default class EqualizerFilter extends Filter
      * Band at 125 Hz
      * @name PIXI.sound.filters.EqualizerFilter.F125
      * @type {Number}
-     * @readOnly
+     * @readonly
      */
     public static F125:number = 125;
     
@@ -53,7 +53,7 @@ export default class EqualizerFilter extends Filter
      * Band at 250 Hz
      * @name PIXI.sound.filters.EqualizerFilter.F250
      * @type {Number}
-     * @readOnly
+     * @readonly
      */
     public static F250:number = 250;
     
@@ -61,7 +61,7 @@ export default class EqualizerFilter extends Filter
      * Band at 500 Hz
      * @name PIXI.sound.filters.EqualizerFilter.F500
      * @type {Number}
-     * @readOnly
+     * @readonly
      */
     public static F500:number = 500;
     
@@ -69,7 +69,7 @@ export default class EqualizerFilter extends Filter
      * Band at 1000 Hz
      * @name PIXI.sound.filters.EqualizerFilter.F1K
      * @type {Number}
-     * @readOnly
+     * @readonly
      */
     public static F1K:number = 1000;
     
@@ -77,7 +77,7 @@ export default class EqualizerFilter extends Filter
      * Band at 2000 Hz
      * @name PIXI.sound.filters.EqualizerFilter.F2K
      * @type {Number}
-     * @readOnly
+     * @readonly
      */
     public static F2K:number = 2000;
     
@@ -85,7 +85,7 @@ export default class EqualizerFilter extends Filter
      * Band at 4000 Hz
      * @name PIXI.sound.filters.EqualizerFilter.F4K
      * @type {Number}
-     * @readOnly
+     * @readonly
      */
     public static F4K:number = 4000;
     
@@ -93,7 +93,7 @@ export default class EqualizerFilter extends Filter
      * Band at 8000 Hz
      * @name PIXI.sound.filters.EqualizerFilter.F8K
      * @type {Number}
-     * @readOnly
+     * @readonly
      */
     public static F8K:number = 8000;
     
@@ -101,7 +101,7 @@ export default class EqualizerFilter extends Filter
      * Band at 16000 Hz
      * @name PIXI.sound.filters.EqualizerFilter.F16K
      * @type {Number}
-     * @readOnly
+     * @readonly
      */
     public static F16K:number = 16000;
 
@@ -109,7 +109,7 @@ export default class EqualizerFilter extends Filter
      * The list of bands 
      * @name PIXI.sounds.filters.EqualizerFilter#bands
      * @type {BiquadFilterNode[]}
-     * @readOnly
+     * @readonly
      */
     public bands:BiquadFilterNode[];
 
@@ -117,7 +117,7 @@ export default class EqualizerFilter extends Filter
      * The map of bands to frequency
      * @name PIXI.sounds.filters.EqualizerFilter#bandsMap
      * @type {Object}
-     * @readOnly
+     * @readonly
      */
     public bandsMap:{[id:number]:BiquadFilterNode};
