+++ src/legacy/index.ts
@@ -1,0 +1,11 @@
+import LegacySound from "./LegacySound";
+import LegacySoundInstance from "./LegacySoundInstance";
+
+/**
+ * Classes supporting non-WebAudio based browsers.
+ * @namespace PIXI.sound.legacy
+ */
+export {
+    LegacySound,
+    LegacySoundInstance
+};
